Refactoring: sudden changes on hierarchy of modules
