+++ game/cli/__init__.py
@@ -1,0 +1,4 @@
+#__init__.py
+#from .skuska import 
+from cli.cli_cards import *
+from cli.cli import *
