+++ game/player/__init__.py
@@ -1,0 +1,2 @@
+from .player import Player
+from .easybot_player import EasyBot
